requirements.txt
openai>=1.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
tiktoken>=0.7.0
transformers>=4.40.0
datasets>=2.18.0
peft>=0.10.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
torch>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
tqdm>=4.66.0
python-dotenv>=1.0.0
